requirements.txt
jinja2==3.1.6
openai==2.4.0
pandas==2.3.3
pillow==12.0.0
python-dotenv==1.1.1
tiktoken==0.12.0